Fixed wdb EOD failure caused by syms containing special chars or leading/trailing spaces.
When dealing with two or more syms which are identical apart from
 special chars or leading/trailing spaces, .wdb.upserttopartition
groups them all into the same sym subpartition in the wdbhdb. This
 can lead to a u-fail when the `p attr cannot be applied to the
 mixed syms. Now, however, .wdb.merge checks whether `p# can be
applied to the contents of each subpartition before they are saved
to the hdb. If not, the data is reordered by the subpartition
column (usually sym), which prevents the EOD failure.