Update dataaccess.q
adjust attributes routing to work with int partitions (note int partitions are created as of period start and so need to define the range of valid values for the last partition of the hdb and current rdb partition, else when querying for data in that partition ( which is > partition value) api will think data is outside the list of partitions.

Alter logic of adjustquery to get the list of partitions from the attributes rather than generate a list incrementally (by +1) from first partition. Allows queries for int partitions to be adjusted.